feat: Add strava club for tournaments
Strava OAUTH2 authorization has a quota of limit 1. The app and
its idea needs to be approved by Strava to get a quota increase
even then the quota would be increased to 999. Hence, the online
events that have strava as data source would take in the
activities of participants via strava clubs. As the admin would
be authorized with the application, the user would be able to
fetch all the activities that took place in the club and manage
the leaderboard